mybatis mapping bug fixed

+++ src/main/java/com/chxyz/demo/dao/UserMapper.java
@@ -2,16 +2,17 @@ package com.chxyz.demo.dao;
 
 import com.chxyz.demo.model.UserDO;
 
+import java.io.Serializable;
 import java.util.List;
 
 
-public interface UserMapper {
+public interface UserMapper  {
 
     List<UserDO> queryAll();
-    UserDO queryUserById(Integer userId);
+    UserDO queryUserById(Long userId);
     UserDO queryUserByName(String userName);
     int insertUser(UserDO userDO);
     int updateUser(UserDO userDO);
-    int deleteUser(Integer userId);
+    int deleteUser(Long userId);
 
 }
