+++ src/main/java/com/chxyz/demo/service/UserService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserService {
 
     List<UserDO> queryAll();
-    UserDO queryUserById(Integer userId);
+    UserDO queryUserById(Long userId);
     UserDO queryUserByName(String userName);
     boolean insertUser(UserDO userDO);
     boolean updateUser(UserDO userDO);
-    boolean deleteUser(Integer userId);
+    boolean deleteUser(Long userId);
 }
